+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.secondproject",
+    package = "com.trucklo",
 )
 
 android_resource(
     name = "res",
-    package = "com.secondproject",
+    package = "com.trucklo",
     res = "src/main/res",
 )
