+++ android/app/src/main/java/com/trucklo/MainActivity.java
@@ -1,4 +1,4 @@
-package com.secondproject;
+package com.trucklo;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "secondProject";
+        return "TruckLo";
     }
 }
