+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'secondProject'
+rootProject.name = 'TruckLo'
 
 include ':app'
 include ':react-native-maps'
