+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "secondProject",
-  "displayName": "secondProject"
+  "name": "TruckLo",
+  "displayName": "TruckLo"
 }
