+++ ios/Podfile
@@ -1,7 +1,7 @@
 # Uncomment the next line to define a global platform for your project
 # platform :ios, '9.0'
 
-target 'secondProject' do
+target 'TruckLo' do
     rn_path = '../node_modules/react-native'
     rn_maps_path = '../node_modules/react-native-maps'
