+++ ios/TruckLo-tvOS/Info.plist
