+++ ios/TruckLo/Images.xcassets/AppIcon.appiconset/Contents.json
