+++ ios/TruckLo/Images.xcassets/Contents.json
