+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "secondProject",
+  "name": "TruckLo",
   "version": "0.0.1",
   "private": true,
   "scripts": {
